fix/frontend: Renamed folder name to avoid case insensitive issues
